Allow content and title to be used for the message
The content element contains the more interesting stuff, but may contain
markup which we remove by using BeautifulSoup get_text() method.

A new configuration variable has been defined to choose between title
and content. This should actually be possible per feed.

The calculation of the length of the message has been changed a bit to
make it more readable.